mikrotik-config-tool/requirements.txt
routeros-api
click
python-dotenv